requirements.txt
nibabel==3.2.1
numpy==1.21.0
packaging==20.9
Pillow==8.3.0
pyparsing==2.4.7
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
typing-extensions==3.10.0.0
